+++ packages/thumbprint-react/icons/index.jsx
@@ -3,3 +3,6 @@ export { default as NavigationCaretDownSmall } from './navigation-caret-down-sma
 export { default as NavigationCloseTiny } from './navigation-close-tiny.jsx';
 export { default as NavigationCloseSmall } from './navigation-close-small.jsx';
 export { default as NavigationCloseMedium } from './navigation-close-medium.jsx';
+export { default as AlertCheck } from './alert-check.jsx';
+export { default as AlertInfo } from './alert-info.jsx';
+export { default as AlertWarning } from './alert-warning.jsx';