+++ packages/thumbprint-react/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 -   [Patch] Update Type syntax for `Image` and `ServiceCardImage` to fix missing API docs.
 
+### Added
+
+-   [Minor] Create `Alert` component in React (TypeScript).
+
 ## 10.0.1 - 2019-10-21
 
 ### Fixed